se agregaron modales para gestionar la actividad de los ususuarios

+++ Js/mainEditarViaje.js
@@ -27,18 +27,51 @@ if (!form) {
   form.addEventListener("submit", (e) => {
     e.preventDefault();
 
-    campos.forEach((campo) => {
-      viaje[campo] = form[campo].type === "number"
-        ? parseFloat(form[campo].value)
-        : form[campo].value;
-    });
+    Swal.fire({
+      title: '¿Querés cambiar el contenido del viaje?',
+      icon: 'question',
+      showCancelButton: true,
+      confirmButtonText: 'Sí, cambiar',
+      cancelButtonText: 'No, cancelar',
+      reverseButtons: true,
+      customClass: {
+        confirmButton: 'btn-confirmar-personalizado'
+      }
+    }).then((result) => {
+      if (result.isConfirmed) {
+        campos.forEach((campo) => {
+          viaje[campo] = form[campo].type === "number"
+            ? parseFloat(form[campo].value)
+            : form[campo].value;
+        });
 
-    const index = viajes.findIndex((v) => v.id === id);
-    viajes[index] = viaje;
+        const index = viajes.findIndex((v) => v.id === id);
+        viajes[index] = viaje;
 
-    localStorage.setItem("viajes", JSON.stringify(viajes));
-    alert("Viaje actualizado correctamente.");
-    window.location.href = "./soyConductor.html";
-  });
+        localStorage.setItem("viajes", JSON.stringify(viajes));
+
+        Swal.fire({
+          title: '¡Listo!',
+          text: 'El viaje fue actualizado correctamente.',
+          icon: 'success',
+          confirmButtonText: 'Volver',
+          customClass: {
+            confirmButton: 'mi-boton-personalizado'
+          },
+          buttonsStyling: false
+        }).then(() => {
+          window.location.href = "./soyConductor.html";
+        });
+      }
+    }
+  )
 }
 
+)
+}
+
+const btnVolver = document.getElementById("btnVolver");
+
+btnVolver.addEventListener("click", () => {
+  window.location.href = "./soyConductor.html";
+});
